requirements.txt
absl-py==1.3.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
contourpy==1.0.5
cycler==0.11.0
faiss-gpu==1.7.2
fonttools==4.38.0
google-auth==2.13.0
google-auth-oauthlib==0.4.6
grpcio==1.50.0
idna==3.4
joblib==1.2.0
kiwisolver==1.4.4
llvmlite==0.39.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.1
numba==0.56.3
numpy==1.23.4
oauthlib==3.2.2
opencv-python==4.6.0.66
packaging==21.3
pandas==1.5.1
Pillow==9.2.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pynndescent==0.5.7
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-metric-learning==1.6.2
pytz==2022.5
record-keeper==0.9.32
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.3
six==1.16.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
torch==1.12.0+cu116
torchvision==0.13.0+cu116
tqdm==4.64.1
typing_extensions==4.4.0
umap-learn==0.5.3
urllib3==1.26.12
Werkzeug==2.2.2
